clean up PatientAppointmentDataGrid
